requirements.txt
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
docker==5.0.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
msgpack-python==0.5.6
mysql-connector==2.2.9
mysqlclient==2.0.3
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-http-client==3.3.2
PyYAML==5.3.1
queuelib==1.5.0
requests==2.25.1
retrying==1.3.3
scrapinghub==2.3.1
Scrapy==2.5.0
sendgrid==6.6.0
service-identity==18.1.0
shub==2.13.0
six==1.15.0
starkbank-ecdsa==1.1.0
toml==0.10.2
tqdm==4.55.1
Twisted==21.2.0
urllib3==1.26.4
w3lib==1.22.0
websocket-client==0.58.0
zope.interface==5.3.0
